Allow configuring TypeReference target instance creator
