+++ Jint.Tests.PublicInterface/Jint.Tests.PublicInterface.csproj
@@ -23,7 +23,7 @@
   <ItemGroup>
     <PackageReference Include="Flurl.Http.Signed" Version="3.2.0" />
     <PackageReference Include="GitHubActionsTestLogger" Version="2.0.1" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
     <PackageReference Include="MongoDB.Bson.signed" Version="2.14.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="xunit" Version="2.4.2" />
